refactor(electron): extract stageRuntimeAssets into a shared module
Moves the runtime-asset staging out of start.action.mjs into
client/electron/stage_runtime_assets.mjs so other launchers (e.g. the
Electron E2E suite) can stage the same assets without duplicating the
list. No behavior change.
